Split generic functionality out of nexrad.
